# Conflicts:
#	code/_onclick/hud/_defines.dm
#	code/_onclick/hud/fullscreen.dm
#	code/_onclick/hud/hud.dm
#	code/_onclick/hud/skybox.dm
#	code/modules/admin/verbs/buildmode.dm
#	code/modules/rigsuits/_rig.dm
#	interface/skin.dmf

+++ code/game/objects/items/weapons/material/material_armor.dm
@@ -75,7 +75,7 @@ Protectiveness | Armor %
 
 	if(istype(source, /obj/item/projectile))
 		var/obj/item/projectile/P = source
-		if(P.pass_flags & PASSGLASS)
+		if(P.check_pass_flags(ATOM_PASS_GLASS))
 			if(material.opacity - 0.3 <= 0)
 				return // Lasers ignore 'fully' transparent material.
 
@@ -94,7 +94,6 @@ Protectiveness | Armor %
 	T.visible_message("<span class='danger'>\The [src] [material.destruction_desc]!</span>")
 	if(istype(loc, /mob/living))
 		var/mob/living/M = loc
-		M.drop_from_inventory(src)
 		if(material.shard_type == SHARD_SHARD) // Wearing glass armor is a bad idea.
 			var/obj/item/material/shard/S = material.place_shard(T)
 			M.embed(S)
@@ -182,7 +181,7 @@ Protectiveness | Armor %
 
 		// Makes sure the numbers stay capped.
 		for(var/number in list(melee_armor, bullet_armor, laser_armor, energy_armor, bomb_armor))
-			number = between(0, number, 100)
+			number = clamp( number, 0,  100)
 
 		armor["melee"] = melee_armor
 		armor["bullet"] = bullet_armor
@@ -191,12 +190,12 @@ Protectiveness | Armor %
 		armor["bomb"] = bomb_armor
 
 		if(!isnull(material.conductivity))
-			siemens_coefficient = between(0, material.conductivity / 10, 10)
+			siemens_coefficient = clamp( material.conductivity / 10, 0,  10)
 		slowdown = clamp(0, round(material.weight / 10, 0.1) * material_weight_factor, 6)
 
 /obj/item/clothing/suit/armor/material
 	name = "armor"
-	default_material = DEFAULT_WALL_MATERIAL
+	default_material = MAT_STEEL
 
 /obj/item/clothing/suit/armor/material/makeshift
 	name = "sheet armor"
@@ -242,9 +241,11 @@ Protectiveness | Armor %
 		if(second_plate.material != src.material)
 			to_chat(user, "<span class='warning'>Both plates need to be the same type of material.</span>")
 			return
-		user.drop_from_inventory(src)
-		user.drop_from_inventory(second_plate)
-		var/obj/item/clothing/suit/armor/material/makeshift/new_armor = new(null, src.material.name)
+		if(!user.attempt_void_item_for_installation(src))
+			return
+		if(!user.attempt_void_item_for_installation(second_plate))
+			return
+		var/obj/item/clothing/suit/armor/material/makeshift/new_armor = new(null, material.name)
 		user.put_in_hands(new_armor)
 		qdel(second_plate)
 		qdel(src)
@@ -271,7 +272,6 @@ Protectiveness | Armor %
 			to_chat(user, "<span class='notice'>You apply some [S.material.use_name] to \the [src].  Hopefully it'll make the makeshift helmet stronger.</span>")
 			var/obj/item/clothing/head/helmet/material/makeshift/helmet = new(null, S.material.name)
 			user.put_in_hands(helmet)
-			user.drop_from_inventory(src)
 			qdel(src)
 			return
 		else
@@ -282,7 +282,7 @@ Protectiveness | Armor %
 /obj/item/clothing/head/helmet/material
 	name = "helmet"
 	flags_inv = HIDEEARS|HIDEEYES|BLOCKHAIR
-	default_material = DEFAULT_WALL_MATERIAL
+	default_material = MAT_STEEL
 
 /obj/item/clothing/head/helmet/material/makeshift
 	name = "bucket"
