+++ code/modules/reagents/reagent_containers/pill.dm
@@ -33,9 +33,10 @@
 			if(blocked)
 				to_chat(user, "<span class='warning'>\The [blocked] is in the way!</span>")
 				return
+			if(!user.attempt_void_item_for_installation(src))
+				return
 
 			to_chat(M, "<span class='notice'>You swallow \the [src].</span>")
-			M.drop_from_inventory(src) //icon update
 			if(reagents.total_volume)
 				reagents.trans_to_mob(M, reagents.total_volume, CHEM_INGEST)
 			qdel(src)
@@ -57,19 +58,15 @@
 		user.setClickCooldown(user.get_attack_speed(src))
 		if(!do_mob(user, M))
 			return
-
-		user.drop_from_inventory(src) //icon update
+		if(!user.attempt_void_item_for_installation(src))
+			return
 		user.visible_message("<span class='warning'>[user] forces [M] to swallow \the [src].</span>")
-
 		var/contained = reagentlist()
 		add_attack_logs(user,M,"Fed a pill containing [contained]")
-
 		if(reagents && reagents.total_volume)
 			reagents.trans_to_mob(M, reagents.total_volume, CHEM_INGEST)
 		qdel(src)
-
 		return 1
-
 	return 0
 
 /obj/item/reagent_containers/pill/afterattack(obj/target, mob/user, proximity)
@@ -305,10 +302,10 @@
 
 /obj/item/reagent_containers/pill/zoom/Initialize(mapload)
 	. = ..()
-	if(prob(50))						//VOREStation edit begin: Zoom pill adjustments
+	if(prob(50))
 		reagents.add_reagent("mold", 2)	//Chance to be more dangerous
 	reagents.add_reagent("expired_medicine", 5)
-	reagents.add_reagent("stimm", 5)	//VOREStation edit end: Zoom pill adjustments
+	reagents.add_reagent("stimm", 5)
 	color = reagents.get_color()
 
 /obj/item/reagent_containers/pill/diet
@@ -318,5 +315,5 @@
 
 /obj/item/reagent_containers/pill/diet/Initialize(mapload)
 	. = ..()
-	reagents.add_reagent("lipozine", 15) //VOREStation Edit
+	reagents.add_reagent("lipozine", 15)
 	color = reagents.get_color()
